# openapi.yaml
openapi: 3.1.0
info:
  title: Rewardz API
  description: |
    OpenAPI 3.1 surface for the 10 endpoints introduced or changed by
    TODO-0018 Colosseum Rewardz League. Nine of the ten endpoints are
    served by the Fastify `api/` service under the `/v1` prefix; the
    tenth (`GET /rewardz/proof/:authority`) is served by the
    `mvp-keeper-bot` Rust service and is documented here so SDK +
    mobile + mini-app consumers can work from a single contract.

    # Scope

    This spec documents ONLY the surface changed by TODO-0018. It is
    not a complete description of every route in the API; pre-existing
    endpoints (points balance, auth, quests GET, admin, etc.) are
    intentionally omitted to keep review tractable. New or
    behaviour-changed endpoints live here with full request/response
    schemas.

    # Contract pinning

    The response field names and HTTP status codes in this spec are
    the contract the SDK (`sdk/packages/sdk`) and mobile app consume.
    Any field rename in the underlying route handler MUST land here in
    lockstep or the SDK round-trip tests break.

    # Keeper vs API

    - Base URL for nine endpoints: `https://api.rewardz.example` (path
      prefix `/v1`).
    - Base URL for the keeper-served proof endpoint:
      `https://keeper.rewardz.example` (no `/v1` prefix; the keeper is
      a separate process with its own Axum router).

  version: 0.1.0-todo18
  contact:
    name: Rewardz team
    url: https://github.com/anthropics/rewardz
servers:
...
paths:
...
  /v1/airdrop/signup:
    post:
      tags: [airdrop]
      summary: Store an email for airdrop notification (encrypted at rest)
      description: |
        Stores an email tied to a wallet for airdrop notification. The
        email is encrypted at rest via pgcrypto
        `pgp_sym_encrypt(email, AIRDROP_EMAIL_KEY)` — plaintext never
        touches disk.

        # Idempotency

        UNIQUE(wallet) + ON CONFLICT DO UPDATE so re-submitting with a
        corrected email overwrites. Last-write-wins is deliberate — a
        user changing their email before distribution is the expected
        case.

        # Fail-closed on missing key

        If `AIRDROP_EMAIL_KEY` is not configured, the endpoint returns
        503 rather than silently storing plaintext.

        Rate-limited to 10 req/minute per IP.
      operationId: airdropSignup
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AirdropSignupRequest"
      responses:
        "200":
          description: Signup stored
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AirdropSignupResponse"
        "400":
          description: Malformed body
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate-limited (10 req/minute per IP)
          headers:
            Retry-After:
              schema:
                type: integer
                minimum: 0
              description: Seconds until the rate-limit window resets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitedError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "503":
          description: AIRDROP_EMAIL_KEY not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ───────────────────────────────────────────────────────────────
  # 5. GET /v1/leaderboard
  # ───────────────────────────────────────────────────────────────
...
  /v1/discovery/featured:
    get:
      tags: [discovery]
      summary: Default discovery tiles for mini-app home + mobile discovery
      description: |
        Default tiles surface for the mini-app home and mobile
        discovery screen.

        # Ranking

        1. `hidden` protocols are excluded entirely.
        2. `at_risk` protocols are kept but dampened to 50% of their
           quality_score — matches the ranking-engine's
           AT_RISK_DAMPENER so dashboards, `/intents/resolve`, and
           this surface all sink at-risk consistently.
        3. Protocols featured in the latest leaderboard snapshot get
           a small additive boost (+0.1) so yesterday's winners
           surface first even when a higher-quality_score protocol
           exists.
        4. Ties broken by creation order (deterministic).

        `rank` is null when the protocol isn't present in today's
        leaderboard snapshot — the tile is still eligible if its
        quality_score pulls it into the top-N.
      operationId: getDiscoveryFeatured
      security: []
      parameters:
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 6
          description: Max tiles to return
      responses:
        "200":
          description: Featured tiles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DiscoveryFeaturedResponse"
        "400":
          description: Malformed query parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          description: Rate-limited (120 req/minute per IP)
          headers:
            Retry-After:
              schema:
                type: integer
                minimum: 0
              description: Seconds until the rate-limit window resets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RateLimitedError"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"

  # ───────────────────────────────────────────────────────────────
  # 7. GET /rewardz/proof/{authority}  (KEEPER-SERVED, no /v1 prefix)
  # ───────────────────────────────────────────────────────────────
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: signed-message
      description: |
        Protocol-owner bearer token obtained from a signed-message
        auth flow. Enforced by `requireBearerAuth +
        requireProtocolOwner` middleware on protocol-scoped admin
        endpoints.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: |
        Protocol API key (prefix `rwz_`). Generated at
        POST /v1/protocols/register. Used by server-to-server
        integrations calling `/v1/points/award`.
    walletAuth:
      type: apiKey
      in: header
      name: X-Wallet-Signature
      description: |
        Signed-message wallet auth. The exact header layout is
        enforced by `requireWalletAuth` middleware and includes the
        base58 pubkey, signature, and a replay-protection nonce.
  schemas:
...
    Error:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          description: Short error code (e.g. "Bad Request", "Not Found")
        message:
          type: string
          description: Human-readable diagnostic
      example:
        error: Bad Request
        message: "Invalid body: wallet must be a valid base58 Solana pubkey"
...
    RateLimitedError:
      type: object
      required: [error, message, retry_after]
      properties:
        error:
          type: string
          enum: ["Too Many Requests"]
        message:
          type: string
        retry_after:
          type: integer
          minimum: 0
          description: |
            Seconds until the rate-limit window resets. Matches the
            `Retry-After` HTTP header emitted alongside the body.
      example:
        error: "Too Many Requests"
        message: "Rate limit exceeded. Try again in 12s."
        retry_after: 12
...
    Base58Pubkey:
      type: string
      pattern: "^[1-9A-HJ-NP-Za-km-z]{32,44}$"
      description: Base58-encoded Solana public key
...
    AirdropSignupRequest:
      type: object
      required: [wallet, email]
      properties:
        wallet:
          $ref: "#/components/schemas/Base58Pubkey"
        email:
          type: string
          format: email
          maxLength: 254
    AirdropSignupResponse:
      type: object
      required: [ok, updated]
      properties:
        ok:
          type: boolean
          description: |
            Always `true` on a 2xx response — the handler only emits
            this shape after a successful write. Failure modes surface
            as 4xx/5xx with the shared `Error` envelope, never as
            `ok: false`.
        updated:
          type: boolean
          description: true = overwrote existing email for this wallet; false = new row

    # ── Leaderboard ───────────────────────────────────────────
...
    DiscoveryFeaturedResponse:
      type: object
      required: [tiles]
      properties:
        tiles:
          type: array
          items:
            $ref: "#/components/schemas/DiscoveryTile"
    DiscoveryTile:
      type: object
      required:
        - protocolId
        - adminWallet
        - qualityScore
        - visibility
        - featured
        - rank
        - referralCode
      properties:
        protocolId:
          type: string
          format: uuid
        adminWallet:
          $ref: "#/components/schemas/Base58Pubkey"
        qualityScore:
          type: [number, "null"]
          format: float
          minimum: 0
          maximum: 1
        visibility:
          type: string
          enum: [active, at_risk]
        featured:
          type: boolean
          description: |
            Whether the protocol appears in the latest leaderboard
            snapshot. Computed server-side as
            `COALESCE(today_rows.featured, false)` — protocols absent
            from today's snapshot are emitted as `false`, never null.
        rank:
          type: [integer, "null"]
          minimum: 1
          description: Rank in today's leaderboard (null if not ranked)
        referralCode:
          type: [string, "null"]

    # ── Rewardz Proof (keeper-served) ─────────────────────────
...